Add getaddrinfo call to convert IP addresses to sockaddr structure
General and former way to build sockaddr structure is to use the Python
tuple `(address, port)`.  However, with IPv6, there is situations not
compatible with this way (e.g., a link-local address such as
"fe80::7%iface").

Using getaddrinfo, the system is responsible of the conversion, we do
not have to worry on it.

Closes #34